chore: build specific cli for k8s
