+++ agent-control/Cargo.toml
@@ -119,8 +119,8 @@ name = "newrelic-config-migrate"
 path = "src/bin/main_config_migrate.rs"
 
 [[bin]]
-name = "newrelic-agent-control-cli"
-path = "src/bin/main_agent_control_cli.rs"
+name = "newrelic-agent-control-k8s-cli"
+path = "src/bin/main_agent_control_k8s_cli.rs"
 
 [features]
 # feature tokio-console allows debugging tokio tasks using tokio-console
