+++ agent-control/src/cli/k8s.rs
@@ -1,0 +1,4 @@
+pub mod errors;
+pub mod install;
+pub mod uninstall;
+pub mod utils;
