+++ agent-control/src/cli/k8s/uninstall/agent_control.rs
@@ -1,10 +1,10 @@
 use crate::agent_control::config::{
     default_group_version_kinds, helmrelease_v2_type_meta, helmrepository_type_meta,
 };
-use crate::cli::errors::CliError;
-use crate::cli::install::agent_control::REPOSITORY_NAME;
-use crate::cli::uninstall::Deleter;
-use crate::cli::utils::try_new_k8s_client;
+use crate::cli::k8s::errors::CliError;
+use crate::cli::k8s::install::agent_control::REPOSITORY_NAME;
+use crate::cli::k8s::uninstall::Deleter;
+use crate::cli::k8s::utils::try_new_k8s_client;
 #[cfg_attr(test, mockall_double::double)]
 use crate::k8s::client::SyncK8sClient;
 use crate::k8s::labels::Labels;
