+++ agent-control/tests/k8s/agent_control_cli/uninstallation.rs
@@ -122,7 +122,7 @@ fn k8s_cli_uninstall_agent_control_clean_empty_cluster() {
 
 /// Builds an uninstallation command for testing purposes with a curated set of defaults and the provided arguments.
 fn ac_uninstall_cmd(namespace: &str, namespace_agents: &str, release_name: &str) -> Command {
-    let mut cmd = Command::cargo_bin("newrelic-agent-control-cli").unwrap();
+    let mut cmd = Command::cargo_bin("newrelic-agent-control-k8s-cli").unwrap();
     cmd.arg("uninstall-agent-control");
     cmd.arg("--namespace").arg(namespace);
     cmd.arg("--namespace-agents").arg(namespace_agents);
